create utils folder for styles
