+++ next.config.ts
@@ -1,8 +1,8 @@
 import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
-  /* config options here */
   reactStrictMode: true,
+  output: "export", // Mengaktifkan static export
 };
 
 export default nextConfig;
