+++ package.json
@@ -31,6 +31,7 @@
     "framer-motion": "^11.11.17",
     "glob": "^11.0.0",
     "google-map-react": "^2.2.1",
+    "input-otp": "^1.4.1",
     "install": "^0.13.0",
     "lucide-react": "^0.461.0",
     "next": "15.0.3",
