--- src/pages/driver/style.ts
@@ -1,11 +1,0 @@
-const useStyles = () => {
-    return {
-        //card box
-        cardBox: "p-6 bg-white shadow-lg rounded-lg border border-gray-200 w-full flex justify-between",
-        
-        //detail transaksi
-        detailTransaksi: "text-2xl font-semibold mb-6 text-white text-center",
-    };
-};
-
-export default useStyles;
