--- src/pages/input_product/style.ts
@@ -1,14 +1,0 @@
-const useStyles = () => {
-    return {
-        //box Input
-        inputBackground: "flex justify-center items-center min-h-screen bg-gray-50",
-        inputEnter: "w-full max-w-md bg-white rounded-lg shadow-lg p-6 space-y-6 my-10",
-        inputAddProduct: "text-2xl font-semibold text-gray-800 text-center",
-
-        //nama produk
-        labelName:"text-gray-700",
-        
-    }
-}
-
-export default useStyles;
