--- src/pages/list_product/style.ts
@@ -1,24 +1,0 @@
-const useStyles = () => {
-    return {
-        // judul
-        containerTittle: "container mx-auto py-10",
-        tittle: "text-2xl font-bold mb-6",
-
-        // list space
-        listSpace: "space-y-6",
-
-        // card
-        card: "flex items-center gap-4 p-4 shadow-md border rounded-lg",
-        cardImage: "w-24 h-24 object-cover rounded-md",
-        cardContent: "flex-1",
-        cardTitle: "text-lg font-semibold",
-        cardDescription: "text-gray-500 text-sm mb-2",
-        cardPrice: "text-gray-700 mb-2",
-        cardFont: "font-medium",
-
-        // card stock
-        cardStock: "text-green-600 font-medium mb-2", // Add styling for stock
-    };
-};
-
-export default useStyles;
