+++ src/pages/login/index.tsx
@@ -8,7 +8,7 @@ import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { Checkbox } from '@/components/ui/checkbox'
 import { Eye, EyeOff, Mail, Facebook } from 'lucide-react'
-import useStyles  from './style'
+import useStyles  from '../../utils/loginStyles'
 import classNames from 'classnames'
 import Navbar from '@/components/ui/navbar'
 import Link from 'next/link'
